Analysis of a Star Wars survey
In this project we are analyzing the answers to a survey from Star Wars fans.
The Survey asks them if they watched all of the movies or just some of them; if they consider themselves as fans or not; what are their favourite characters and in what order and finally some personal datas are asked. The survey remains totally anonimous.